Work around DelSolr's moronitude.
DelSolr builds XML via string concatenation, which makes Baby Cthulhu cry. It
helpfully offers a cdata option that wraps the value in a CDATA block, but this
is repetitive and error-prone. Instead, let's wrap the document in something
that does the right thing by always passing an escaped string to
DelSolr::Document.