Lock down access to create new needs and to full edit page
We were hiding the links to these from non-admin users but not actually checking
what happened when someone manually entered the URLs. We now check that.

We're still not checking what gets posted to the update action to make sure that
the user is only editing things they're allowed to edit. I'm going to treat that
as a separate change.