Add tire gem for search and bump gds-sso
I had to bump gds-sso as some of the old version's dependencies are no longer available.
Tire is a wrapper for Elasticsearch. I'm not a fan of the way it couples search concerns
into the activerecord model, but it is a quick way to get things up and running and my
main concern now is to get need-o-tron off solr since we're standardising on elasticsearch
everywhere else.

In time we should explore whether needs should simply use rummager for search.