Re-enable sorting
In order to sort on title we need a non-indexed version of the title as elasticsearch
won't sort on a field that's been broken into multiple tokens. We do that with a multi
field and some extra code that shapes the search query accordingly. This feels a bit
ugly but gets the job done.